+++ imod/common/interfaces/ilinedatapackage.py
@@ -1,7 +1,7 @@
 from abc import abstractmethod
 from typing import TYPE_CHECKING
 
-from imod.mf6.interfaces.ipackagebase import IPackageBase
+from imod.common.interfaces.ipackagebase import IPackageBase
 
 if TYPE_CHECKING:
     import geopandas as gpd
