+++ imod/common/interfaces/ipackage.py
@@ -2,7 +2,7 @@ import abc
 from abc import abstractmethod
 from typing import Any
 
-from imod.mf6.interfaces.ipackagebase import IPackageBase
+from imod.common.interfaces.ipackagebase import IPackageBase
 
 
 class IPackage(IPackageBase, metaclass=abc.ABCMeta):
