+++ imod/mf6/dis.py
@@ -5,10 +5,10 @@ from typing import Any, List, Optional, Union
 import numpy as np
 
 import imod
+from imod.common.interfaces.imaskingsettings import IMaskingSettings
+from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.logging import init_log_decorator, standard_log_decorator
 from imod.mf6.disv import VerticesDiscretization
-from imod.mf6.interfaces.imaskingsettings import IMaskingSettings
-from imod.mf6.interfaces.iregridpackage import IRegridPackage
 from imod.mf6.package import Package
 from imod.mf6.regrid.regrid_schemes import DiscretizationRegridMethod, RegridMethodType
 from imod.mf6.utilities.grid import create_smallest_target_grid
