+++ imod/mf6/dsp.py
@@ -1,7 +1,7 @@
 import numpy as np
 
+from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.logging import init_log_decorator
-from imod.mf6.interfaces.iregridpackage import IRegridPackage
 from imod.mf6.package import Package
 from imod.mf6.regrid.regrid_schemes import DispersionRegridMethod
 from imod.mf6.validation import PKG_DIMS_SCHEMA
