+++ imod/mf6/multimodel/modelsplitter.py
@@ -2,6 +2,7 @@ from typing import List, NamedTuple
 
 import numpy as np
 
+from imod.common.interfaces.imodel import IModel
 from imod.logging import logger
 from imod.mf6.auxiliary_variables import (
     expand_transient_auxiliary_variables,
@@ -9,7 +10,6 @@ from imod.mf6.auxiliary_variables import (
 )
 from imod.mf6.boundary_condition import BoundaryCondition
 from imod.mf6.hfb import HorizontalFlowBarrierBase
-from imod.mf6.interfaces.imodel import IModel
 from imod.mf6.utilities.clip import clip_by_grid
 from imod.mf6.utilities.grid import get_active_domain_slice
 from imod.mf6.wel import Well
