+++ imod/mf6/npf.py
@@ -5,8 +5,8 @@ from typing import Optional
 import numpy as np
 import xarray as xr
 
+from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.logging import init_log_decorator
-from imod.mf6.interfaces.iregridpackage import IRegridPackage
 from imod.mf6.package import Package
 from imod.mf6.regrid.regrid_schemes import (
     NodePropertyFlowRegridMethod,
