+++ imod/mf6/oc.py
@@ -5,8 +5,8 @@ from typing import Any, Dict, Union
 
 import numpy as np
 
+from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.logging import init_log_decorator
-from imod.mf6.interfaces.iregridpackage import IRegridPackage
 from imod.mf6.package import Package
 from imod.mf6.utilities.dataset import is_dataarray_none
 from imod.mf6.write_context import WriteContext
