+++ imod/mf6/package.py
@@ -14,11 +14,11 @@ import xarray as xr
 import xugrid as xu
 
 import imod
+from imod.common.interfaces.ipackage import IPackage
 from imod.logging import standard_log_decorator
 from imod.mf6.auxiliary_variables import (
     get_variable_names,
 )
-from imod.mf6.interfaces.ipackage import IPackage
 from imod.mf6.pkgbase import (
     EXCHANGE_PACKAGES,
     TRANSPORT_PACKAGES,
