+++ imod/mf6/ssm.py
@@ -1,9 +1,9 @@
 import numpy as np
 
+from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.logging import init_log_decorator, logger
 from imod.mf6 import GroundwaterFlowModel
 from imod.mf6.boundary_condition import BoundaryCondition
-from imod.mf6.interfaces.iregridpackage import IRegridPackage
 from imod.schemata import DTypeSchema
 from imod.util.regrid_method_type import EmptyRegridMethod, RegridMethodType
