+++ imod/msw/initial_conditions.py
@@ -2,7 +2,7 @@ import pathlib
 import shutil
 from typing import Any, TextIO
 
-from imod.mf6.interfaces.iregridpackage import IRegridPackage
+from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.msw.fixed_format import VariableMetaData
 from imod.msw.pkgbase import MetaSwapPackage
 from imod.util.regrid_method_type import EmptyRegridMethod, RegridMethodType
