+++ imod/msw/ponding.py
@@ -3,7 +3,7 @@ from typing import Any, TextIO
 import pandas as pd
 import xarray as xr
 
-from imod.mf6.interfaces.iregridpackage import IRegridPackage
+from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.msw.fixed_format import VariableMetaData
 from imod.msw.pkgbase import DataDictType, MetaSwapPackage
 from imod.msw.regrid.regrid_schemes import PondingRegridMethod
