+++ imod/msw/scaling_factors.py
@@ -1,4 +1,4 @@
-from imod.mf6.interfaces.iregridpackage import IRegridPackage
+from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.msw.fixed_format import VariableMetaData
 from imod.msw.pkgbase import MetaSwapPackage
 from imod.msw.regrid.regrid_schemes import ScalingRegridMethod
