+++ imod/msw/vegetation.py
@@ -3,7 +3,7 @@ from typing import Any, TextIO
 import numpy as np
 import xarray as xr
 
-from imod.mf6.interfaces.iregridpackage import IRegridPackage
+from imod.common.interfaces.iregridpackage import IRegridPackage
 from imod.msw.fixed_format import VariableMetaData
 from imod.msw.pkgbase import MetaSwapPackage
 from imod.util.regrid_method_type import EmptyRegridMethod, RegridMethodType
